module predict_class.py added
