+++ module/train_model.py
@@ -14,7 +14,8 @@ from module.evaluate_model import evaluate_model
 logger = setup_logger("train_model")
 
 
-def train_model(directory: str,
+def train_model(train_dataset,
+                test_dataset,
                 batch_size: int,
                 image_size: tuple[int, int],
                 num_classes: int = 58,
@@ -28,53 +29,24 @@ def train_model(directory: str,
 
       Параметры:
       ----------
-      directory : str
-          Путь к директории, содержащей изображения, организованные по подкаталогам для каждой
-          категории (метки). Директория должна иметь подкаталоги, где каждая папка соответствует
-          отдельному классу для классификации.
 
-      batch_size : int
-          Размер пакета (batch) для загрузки данных. Задает количество изображений, передаваемых
-          в сеть за один шаг, что влияет на скорость и эффективность обучения.
-
-      image_size : tuple[int, int]
-          Размер загружаемых изображений в формате (высота, ширина). Все изображения масштабируются
-          до этого размера, обеспечивая согласованность данных.
-      num_classes : int
-          Количество классов для распознавания
 
       Возвращаемое значение:
       ----------------------
-      None
-          Функция не возвращает значений, но записывает информацию в лог. В ходе работы функция
-          выводит сообщения о начале и успешном завершении обучения. В случае возникновения ошибок
-          информация также записывается в лог.
+
 
       Логика работы:
       --------------
-      1. **Загрузка данных** — функция `load_train_test_val` загружает данные из `directory`, разбивая
-         их на обучающий, тестовый и валидационный наборы данных.
-      2. **Создание модели** — функция `create_model` создает сверточную нейронную сеть, подходящую для
-         указанной задачи классификации, определяя входную форму и количество классов на основе загруженных
-         данных.
-      3. **Компиляция модели** — модель компилируется с использованием стандартных параметров.
-      4. **Обучение модели** — модель обучается с использованием метода `fit` и колбэка `EarlyStopping`,
-         который завершает обучение при отсутствии улучшений, предотвращая переобучение.
-      5. **Оценка модели** — модель оценивается на валидационном наборе с использованием функции `evaluate_model`,
-         чтобы проверить её точность после обучения.
-      6. **Логирование** — начало, успешное завершение и ошибки в процессе обучения записываются в лог.
+
 
       Примечание:
       -----------
       Функция `create_model` определяет количество классов на основе количества поддиректорий в `directory`,
       предполагая, что каждая поддиректория представляет один класс.
       """
-    logger.info(f"training starts with directory: {directory}, batch_size: {batch_size}, image_size: {image_size}")
+    logger.info(f"training starts with batch_size: {batch_size}, image_size: {image_size}, num_classes: {num_classes}")
     try:
-        train_dataset, test_dataset = load_train_test_val(directory, batch_size, image_size)
-        print(train_dataset)
         model = create_model(input_shape=image_size + (3,), num_classes=num_classes)
-
         # Показать модель
         model.summary()
         metrics = [
@@ -102,7 +74,6 @@ def train_model(directory: str,
             verbose=1
 
         )
-        f1_score = []
         # Определите колбек для ранней остановки
         early_stopping_callback = tf.keras.callbacks.EarlyStopping(
             monitor='val_categorical_accuracy',  # Мониторим валидационную потерю
